Pass the AssetConfig and required components down (#212)
* Passes the AssetConfig and required components down

When using assets which are generically matched, such as **/*.scss, the processor associated with this asset is used for all matching assets. This means the asset stored in the processor is the generic one.

Because the asset in the processor is the generic one, the processor is unable to know specific details about the non-generic asset. Passing these values down the chain allows the processor to access this asset-specific information in a generic context.

Other files affected by this change have been updated to reflect it.

* Fix associated tests and sample