Fix for change in .NET 6 for context.Request.Path.Value (#240)
* Fix for change in .NET 6 for context.Request.Path.Value
[release]

In .NET 6 the value of context.Request.Path.Value appears to always have a leading "\" regardless of the reference in the html file.  An example of this is when you add a javascript bundle like:
pipeline.AddJavaScriptBundle("Scripts/libraries.js", "Scripts/jquery.gridster.js");
and reference it in the html file like this:
<script src="Scripts/libraries.js"></script>
The bundle is added as an asset with the value of "Scripts/libraries.js" but the value of context.Request.Path.Value will be "/Scripts/libraries.js" resulting in no match.
This proposed change will sanitize all assets to have a leading "\" to ensure this matches.  
In my companies use of multiple javascript bundles with less compilation, this worked for us, but I'm not sure if this will mess with/break other possible routes.  This needs someone who is more familiar with the entirety of the WebOptimizer solution to ensure it doesn't introduce regressions for other people.  (I also tested it in older versions of .NET and it appeared to work without issue)
Hopefully @madskristensen or one of the others can take a look and evaluate the regression possibilities of this one.

* Fixed tests